Add libcam.client wrapper
Before Oreo, the buffer_handle_t passing to camera HAL was from an
ANativeWindowBuffer struct. MTK's libcam.client had a simple size check
when create the imgage buffer. It Assumed the buffer_handle_t was from
an ANativeWindowBuffer struct but not anymore. This check should be done
in the callee now. This wrapper is to avoid the crash in libcam.client.
It only replace the image buffer class.
TODO: It seems that libcam.client is hardware independent, it might be
rebuilt without the wrapper.